Basic implementation of the container stats SDK
This is a very basic and work-in-progress implementation.

I plan to get a webhook and front-end integration PoC working as soon as I can, then I'll put in the details and optimizations later.

There needs to be done a check on Docker's formula for calculating the stats, because there have are some differences from what's returned in the API and what's returned in `docker stats`.

More details have been documented on the @TODO comments.